+++ src/app/PongApp.hpp
@@ -2,6 +2,7 @@
 
 #include "App.hpp"
 #include "PongLogic.hpp"
+#include "../detail/TextSurfaceCache.hpp"
 #include <SDL2/SDL.h>
 #include <SDL2/SDL_ttf.h>
 
@@ -22,6 +23,7 @@ public:
     void onFocusGained() override;
     void onFocusLost() override;
     void onResize(int clientWidth, int clientHeight) override;
+    void onUiScaleChanged() override;
 
 private:
     static constexpr int kHudHeight = 32;
@@ -34,6 +36,7 @@ private:
     int hudHeight() const;
 
     TTF_Font* m_font = nullptr;
+    mutable monolith::detail::TextSurfaceCache m_textSurfaceCache;
     monolith::pong::Game m_game;
     bool m_paused = false;
     bool m_holdUp = false;